refactor: response 인터페이스 공통 폴더로 분리(#81)
