+++ src/types/common/apiResponse.ts
@@ -1,0 +1,7 @@
+export interface ApiResponse<T> {
+  success: boolean;
+  code: string;
+  message: string;
+  timestamp: string;
+  data: T;
+}
